Move more things up to the base Corr2 class
